requirements.txt
diffusers==0.31.0
bitsandbytes==0.43.3
transformers==4.46.3
accelerate==0.34.2
torchvision==0.19.1
wandb
xformers
pytorch-fid
scipy==1.11.1
opencv-python==4.8.0.74
peft==0.13.2
git+https://github.com/openai/CLIP.git
lpips
torchmetrics[image]